+++ stroom-util-shared/src/main/java/stroom/util/shared/PropertyPath.java
@@ -1,5 +1,7 @@
 package stroom.util.shared;
 
+import stroom.docref.SharedObject;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -10,7 +12,9 @@ import java.util.Objects;
  * Class for representing a path to a property in an object tree, i.e
  * stroom.node.name
  */
-public class PropertyPath implements Comparable<PropertyPath> {
+public class PropertyPath implements SharedObject, Comparable<PropertyPath> {
+
+    private static final long serialVersionUID = 4298017138182350850L;
 
     private static final String DELIMITER = ".";
     private static final String DELIMITER_REGEX = "\\" + DELIMITER;
@@ -19,7 +23,15 @@ public class PropertyPath implements Comparable<PropertyPath> {
 
     // Held in part form to reduce memory overhead as some parts will be used
     // many times over all the config objects
-    private final List<String> parts;
+    private List<String> parts;
+
+    @SuppressWarnings("unused") // Needed for GWT
+    private PropertyPath() {
+    }
+
+    private PropertyPath(final List<String> parts) {
+        this.parts = new ArrayList<>(parts);
+    }
 
     public static PropertyPath blank() {
         return EMPTY_INSTANCE;
@@ -48,10 +60,6 @@ public class PropertyPath implements Comparable<PropertyPath> {
         } else {
             return new PropertyPath(parts);
         }
-    }
-
-    private PropertyPath(final List<String> parts) {
-        this.parts = new ArrayList<>(parts);
     }
 
     /**
@@ -123,9 +131,15 @@ public class PropertyPath implements Comparable<PropertyPath> {
         return Objects.hash(parts);
     }
 
-    public static class Builder {
+    public static class Builder implements SharedObject {
+
+        private static final long serialVersionUID = 4646431868659034045L;
 
         private List<String> parts = null;
+
+        @SuppressWarnings("unused") // Needed for GWT
+        private Builder() {
+        }
 
         /**
          * Add path part to the end of the list of paths already added